+++ src/js/_legacy/utils/formatters/index.ts
@@ -1,2 +1,0 @@
-export { formatBibliography } from './format-bibliography';
-export { formatReferenceLinks } from './format-reference-links';