+++ src/js/utils/ajax.ts
@@ -1,12 +1,13 @@
 import { addQueryArgs } from '@wordpress/url';
 
 declare const _abt_nonce: string;
+declare const ajaxurl: string;
 
 export async function fetchAjax(
     action: string,
     body?: Record<string, string | number | boolean>,
 ): Promise<Response> {
-    return fetch(top.ajaxurl, {
+    return fetch(ajaxurl, {
         method: 'POST',
         headers: {
             'Content-type': 'application/x-www-form-urlencoded',