+++ src/js/utils/error.ts
@@ -1,4 +1,4 @@
-type ResponseData = Omit<Response, keyof Body>;
+type ResponseData = Pick<Response, Exclude<keyof Response, keyof Body>>;
 
 export class ResponseError extends Error {
     resource: string;