+++ src/php/editor.php
@@ -13,9 +13,7 @@ defined( 'ABSPATH' ) || exit;
 
 use function ABT\Utils\{
 	add_json_script,
-	get_citation_styles,
 	get_handle,
-	is_block_editor
 };
 
 /**